+++ webapp/CHANGELOG.md
@@ -1,9 +1,8 @@
 ## [0.68.2](https://github.com/SocialGouv/carte-jeune-engage/compare/v0.68.1...v0.68.2) (2024-11-19)
 
-
 ### Bug Fixes
 
-* add defensive code in getserversideprops offer & log ([07d6e1b](https://github.com/SocialGouv/carte-jeune-engage/commit/07d6e1b1a9dd4738d6943558d17fbbd7aa321c85))
+- add defensive code in getserversideprops offer & log ([07d6e1b](https://github.com/SocialGouv/carte-jeune-engage/commit/07d6e1b1a9dd4738d6943558d17fbbd7aa321c85))
 
 ## [0.68.1](https://github.com/SocialGouv/carte-jeune-engage/compare/v0.68.0...v0.68.1) (2024-11-19)